refactor: update SocialMediaSection and SocialMediaPostCarousel components
