+++ packages/app-sundevil/src/components/index.js
@@ -3,4 +3,5 @@ export * from "./Header";
 export * from "./HeaderContentSportLinks";
 export * from "./MyComponent";
 export * from "./OfficialAthleticsSite";
+export * from "./SocialMediaSection";
 //
